requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
black==20.8b1
chex
click==8.0.4
dm-haiku
flake8==3.8.4
flax < 0.5.0
hydra-core==1.2.0
ipython==8.5.0
isort==5.6.4
matplotlib==3.5.2
numpy==1.22.*
optax
pandas==1.5.2
protobuf==3.19.6
redis==3.5.3
scipy==1.8.0
tensorflow-probability==0.15.0
typing-extensions >= 3.10
jsonschema==4.17.3
brax==0.0.12
jumanji
scikit-learn